Updated UI
Introduced new scenes for main menu and options. Buttons are clickable. To start game, open Main_Menu scene, unload all the others, then press start on the unity ui. Buttons work from there on out. Also introduced exit button for when we're in game so we don't need to reload the entire game to start it again, if you fall off the map just go back to the main menu.